frontend/requirements.txt
# streamlit frontend requirements
streamlit
streamlit_authenticator
requests
PyYAML
